+++ tests/common/damm_v2.ts
@@ -162,18 +162,36 @@ export function getDammV2RemainingAccounts(
   return remainingAccounts;
 }
 
-export async function createDammV2Pool(
-  svm: LiteSVM,
-  creator: Keypair,
-  tokenAMint: PublicKey,
-  tokenBMint: PublicKey,
-  amountA?: BN,
-  amountB?: BN,
-  baseFeeParams?: Buffer,
-  compoundingFeeBps?: number,
-  collectFeeMode?: number
-): Promise<PublicKey> {
+export async function createDammV2Pool(params: {
+  svm: LiteSVM;
+  creator: Keypair;
+  tokenAMint: PublicKey;
+  tokenBMint: PublicKey;
+  amountA?: BN;
+  amountB?: BN;
+  baseFee?: Buffer;
+  sqrtMinPrice?: BN;
+  sqrtMaxPrice?: BN;
+  initSqrtPrice?: BN;
+  collectFeeMode?: number;
+  compoundingFeeBps?: number;
+}): Promise<PublicKey> {
+  const {
+    svm,
+    creator,
+    tokenAMint,
+    tokenBMint,
+    amountA,
+    amountB,
+    baseFee: baseFeeParams,
+    collectFeeMode,
+    compoundingFeeBps,
+  } = params;
   const program = createDammV2Program();
+
+  const sqrtMinPrice = params.sqrtMinPrice ?? MIN_SQRT_PRICE;
+  const sqrtMaxPrice = params.sqrtMaxPrice ?? MAX_SQRT_PRICE;
+  const sqrtPrice = params.initSqrtPrice ?? INIT_PRICE;
 
   const poolAuthority = deriveDammV2PoolAuthority();
   const pool = deriveDammV2CustomizablePoolAddress(tokenAMint, tokenBMint);
@@ -207,17 +225,31 @@ export async function createDammV2Pool(
   if (amountA && amountB) {
     const liquidityFromA = getLiquidityDeltaFromAmountA(
       amountA,
-      INIT_PRICE,
-      MAX_SQRT_PRICE
+      sqrtPrice,
+      sqrtMaxPrice
     );
 
     const liquidityFromB = getLiquidityDeltaFromAmountB(
       amountB,
-      MIN_SQRT_PRICE,
-      INIT_PRICE
+      sqrtMinPrice,
+      sqrtPrice
     );
 
     liquidityDelta = BN.min(liquidityFromA, liquidityFromB);
+  } else if (amountA) {
+    // one sided pool A
+    liquidityDelta = getLiquidityDeltaFromAmountA(
+      amountA,
+      sqrtPrice,
+      sqrtMaxPrice
+    );
+  } else if (amountB) {
+    // one sided pool B
+    liquidityDelta = getLiquidityDeltaFromAmountB(
+      amountB,
+      sqrtMinPrice,
+      sqrtPrice
+    );
   }
 
   const baseFee = {
@@ -241,11 +273,11 @@ export async function createDammV2Pool(
         padding: 0,
         dynamicFee: null,
       },
-      sqrtMinPrice: MIN_SQRT_PRICE,
-      sqrtMaxPrice: MAX_SQRT_PRICE,
+      sqrtMinPrice,
+      sqrtMaxPrice,
       hasAlphaVault: false,
       liquidity: liquidityDelta,
-      sqrtPrice: INIT_PRICE,
+      sqrtPrice,
       activationType: 0,
       collectFeeMode: collectFeeMode ?? 1,
       activationPoint: null,
@@ -284,8 +316,12 @@ export async function createDammV2Pool(
 
   const vaultBBalance = Number(AccountLayout.decode(tokenBVaultData).amount);
 
-  expect(vaultABalance).greaterThan(0);
-  expect(vaultBBalance).greaterThan(0);
+  if (!sqrtPrice.eq(sqrtMaxPrice)) {
+    expect(vaultABalance).greaterThan(0);
+  }
+  if (!sqrtPrice.eq(sqrtMinPrice)) {
+    expect(vaultBBalance).greaterThan(0);
+  }
 
   return pool;
 }